refactor: colocate domain SQL tests
